HDDS-8487. Recon Heatmap UI Changes and Size Normalization Ivan Review Comment
