+++ hadoop-ozone/recon/src/main/resources/webapps/recon/ozone-recon-web/src/views/heatMap/heatmap.less
@@ -50,6 +50,7 @@
     fill: rgba(0, 0, 0, 0.65) !important;
     line-height: 1.5;
     font-feature-settings: 'tnum', "tnum";
+    height: 800;
   }
 
 }
